Reorganize demos
* Reorganize demo folder
* Fix out-of-date mmdet/mmtrack configs
* Fix a bug in inference_top_sown_pose_model which causes track_id missing (seems introduced by #560)

Remaining issues:
* Some video files used in example commands in demos do not exist
