+++ demo/2d_face_demo.md
@@ -13,7 +13,7 @@ The pre-trained face keypoint estimation model can be found from [model zoo](htt
 Take [aflw model](https://download.openmmlab.com/mmpose/face/hrnetv2/hrnetv2_w18_aflw_256x256-f2bbc62b_20210125.pth) as an example:
 
 ```shell
-python demo/top_down_img_demo.py \
+python demo/scripts/top_down_img_demo.py \
     ${MMPOSE_CONFIG_FILE} ${MMPOSE_CHECKPOINT_FILE} \
     --img-root ${IMG_ROOT} --json-file ${JSON_FILE} \
     --out-img-root ${OUTPUT_DIR} \
@@ -24,7 +24,7 @@ python demo/top_down_img_demo.py \
 Examples:
 
 ```shell
-python demo/top_down_img_demo.py \
+python demo/scripts/top_down_img_demo.py \
     configs/face/hrnetv2/aflw/hrnetv2_w18_aflw_256x256.py \
     https://download.openmmlab.com/mmpose/face/hrnetv2/hrnetv2_w18_aflw_256x256-f2bbc62b_20210125.pth \
     --img-root tests/data/onehand10k/ --json-file tests/data/onehand10k/test_onehand10k.json \
@@ -34,7 +34,7 @@ python demo/top_down_img_demo.py \
 To run demos on CPU:
 
 ```shell
-python demo/top_down_img_demo.py \
+python demo/scripts/top_down_img_demo.py \
     configs/face/hrnetv2/aflw/hrnetv2_w18_aflw_256x256.py \
     https://download.openmmlab.com/mmpose/face/hrnetv2/hrnetv2_w18_aflw_256x256-f2bbc62b_20210125.pth \
     --img-root tests/data/aflw/ --json-file tests/data/aflw/test_aflw.json \
@@ -50,7 +50,7 @@ Please install `face_recognition` before running the demo, by `pip install face_
 For more details, please refer to https://github.com/ageitgey/face_recognition.
 
 ```shell
-python demo/face_img_demo.py \
+python demo/scripts/face_img_demo.py \
     ${MMPOSE_CONFIG_FILE} ${MMPOSE_CHECKPOINT_FILE} \
     --img-root ${IMG_ROOT} --img ${IMG_FILE} \
     --out-img-root ${OUTPUT_DIR} \
@@ -59,7 +59,7 @@ python demo/face_img_demo.py \
 ```
 
 ```shell
-python demo/face_img_demo.py \
+python demo/scripts/face_img_demo.py \
     configs/face/hrnetv2/aflw/hrnetv2_w18_aflw_256x256.py \
     https://download.openmmlab.com/mmpose/face/hrnetv2/hrnetv2_w18_aflw_256x256-f2bbc62b_20210125.pth \
     --img-root tests/data/aflw/ \
@@ -75,7 +75,7 @@ Please install `face_recognition` before running the demo, by `pip install face_
 For more details, please refer to https://github.com/ageitgey/face_recognition.
 
 ```shell
-python demo/face_video_demo.py \
+python demo/scripts/face_video_demo.py \
     ${MMPOSE_CONFIG_FILE} ${MMPOSE_CHECKPOINT_FILE} \
     --video-path ${VIDEO_FILE} \
     --out-video-root ${OUTPUT_VIDEO_ROOT} \
@@ -86,10 +86,10 @@ python demo/face_video_demo.py \
 Examples:
 
 ```shell
-python demo/face_video_demo.py \
+python demo/scripts/face_video_demo.py \
     configs/face/hrnetv2/aflw/hrnetv2_w18_aflw_256x256.py \
     https://download.openmmlab.com/mmpose/face/hrnetv2/hrnetv2_w18_aflw_256x256-f2bbc62b_20210125.pth \
-    --video-path demo/demo_video.mp4 \
+    --video-path demo/resources/demo_video.mp4 \
     --out-video-root vis_results
 ```
