+++ demo/mmtracking_cfg/tracktor_faster-rcnn_r50_fpn_4e_mot17-private.py
@@ -79,11 +79,9 @@ model = dict(
                 pos_weight=-1,
                 debug=False),
             rpn_proposal=dict(
-                nms_across_levels=False,
                 nms_pre=2000,
-                nms_post=1000,
-                max_num=1000,
-                nms_thr=0.7,
+                max_per_img=1000,
+                nms=dict(type='nms', iou_threshold=0.7),
                 min_bbox_size=0),
             rcnn=dict(
                 assigner=dict(
@@ -103,11 +101,9 @@ model = dict(
                 debug=False)),
         test_cfg=dict(
             rpn=dict(
-                nms_across_levels=False,
                 nms_pre=1000,
-                nms_post=1000,
-                max_num=1000,
-                nms_thr=0.7,
+                max_per_img=1000,
+                nms=dict(type='nms', iou_threshold=0.7),
                 min_bbox_size=0),
             rcnn=dict(
                 score_thr=0.05,
