+++ demo/scripts/bottom_up_img_demo.py
