+++ demo/scripts/top_down_img_demo_with_mmdet.py
