+++ demo/scripts/top_down_pose_tracking_demo_with_mmdet.py
@@ -2,10 +2,38 @@ import os
 from argparse import ArgumentParser
 
 import cv2
-import numpy as np
 
-from mmpose.apis import (inference_top_down_pose_model, init_pose_model,
-                         vis_pose_result)
+from mmpose.apis import (get_track_id, inference_top_down_pose_model,
+                         init_pose_model, vis_pose_tracking_result)
+
+try:
+    from mmdet.apis import inference_detector, init_detector
+    has_mmdet = True
+except (ImportError, ModuleNotFoundError):
+    has_mmdet = False
+
+
+def process_mmdet_results(mmdet_results, cat_id=0):
+    """Process mmdet results, and return a list of bboxes.
+
+    :param mmdet_results:
+    :param cat_id: category id (default: 0 for human)
+    :return: a list of detected bounding boxes
+    """
+    if isinstance(mmdet_results, tuple):
+        det_results = mmdet_results[0]
+    else:
+        det_results = mmdet_results
+
+    bboxes = det_results[cat_id]
+
+    person_results = []
+    for bbox in bboxes:
+        person = {}
+        person['bbox'] = bbox
+        person_results.append(person)
+
+    return person_results
 
 
 def main():
@@ -14,6 +42,8 @@ def main():
     Using mmdet to detect the human.
     """
     parser = ArgumentParser()
+    parser.add_argument('det_config', help='Config file for detection')
+    parser.add_argument('det_checkpoint', help='Checkpoint file for detection')
     parser.add_argument('pose_config', help='Config file for pose')
     parser.add_argument('pose_checkpoint', help='Checkpoint file for pose')
     parser.add_argument('--video-path', type=str, help='Video path')
@@ -30,11 +60,30 @@ def main():
     parser.add_argument(
         '--device', default='cuda:0', help='Device used for inference')
     parser.add_argument(
+        '--bbox-thr',
+        type=float,
+        default=0.3,
+        help='Bounding box score threshold')
+    parser.add_argument(
         '--kpt-thr', type=float, default=0.3, help='Keypoint score threshold')
+    parser.add_argument(
+        '--iou-thr', type=float, default=0.3, help='IoU score threshold')
+    parser.add_argument(
+        '--det-cat-id',
+        type=int,
+        default=0,
+        help='Category id for bounding box detection model')
+
+    assert has_mmdet, 'Please install mmdet to run the demo.'
 
     args = parser.parse_args()
 
     assert args.show or (args.out_video_root != '')
+    assert args.det_config is not None
+    assert args.det_checkpoint is not None
+
+    det_model = init_detector(
+        args.det_config, args.det_checkpoint, device=args.device.lower())
     # build the pose model from a config file and a checkpoint file
     pose_model = init_pose_model(
         args.pose_config, args.pose_checkpoint, device=args.device.lower())
@@ -42,9 +91,7 @@ def main():
     dataset = pose_model.cfg.data['test']['type']
 
     cap = cv2.VideoCapture(args.video_path)
-    fps = cap.get(cv2.CAP_PROP_FPS)
-    size = (int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
-            int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
+    assert cap.isOpened(), f'Faild to load video file {args.video_path}'
 
     if args.out_video_root == '':
         save_out_video = False
@@ -53,6 +100,9 @@ def main():
         save_out_video = True
 
     if save_out_video:
+        fps = cap.get(cv2.CAP_PROP_FPS)
+        size = (int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
+                int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
         fourcc = cv2.VideoWriter_fourcc(*'mp4v')
         videoWriter = cv2.VideoWriter(
             os.path.join(args.out_video_root,
@@ -65,26 +115,37 @@ def main():
     # e.g. use ('backbone', ) to return backbone feature
     output_layer_names = None
 
+    next_id = 0
+    pose_results = []
     while (cap.isOpened()):
+        pose_results_last = pose_results
+
         flag, img = cap.read()
         if not flag:
             break
+        # test a single image, the resulting box is (x1, y1, x2, y2)
+        mmdet_results = inference_detector(det_model, img)
 
         # keep the person class bounding boxes.
-        person_results = [{'bbox': np.array([0, 0, size[0], size[1]])}]
+        person_results = process_mmdet_results(mmdet_results, args.det_cat_id)
 
         # test a single image, with a list of bboxes.
         pose_results, returned_outputs = inference_top_down_pose_model(
             pose_model,
             img,
             person_results,
+            bbox_thr=args.bbox_thr,
             format='xyxy',
             dataset=dataset,
             return_heatmap=return_heatmap,
             outputs=output_layer_names)
 
+        # get track id for each person instance
+        pose_results, next_id = get_track_id(
+            pose_results, pose_results_last, next_id, iou_thr=args.iou_thr)
+
         # show the results
-        vis_img = vis_pose_result(
+        vis_img = vis_pose_tracking_result(
             pose_model,
             img,
             pose_results,
