+++ demo/scripts/top_down_video_demo_full_frame_without_det.py
@@ -2,31 +2,10 @@ import os
 from argparse import ArgumentParser
 
 import cv2
+import numpy as np
 
 from mmpose.apis import (inference_top_down_pose_model, init_pose_model,
-                         vis_pose_tracking_result)
-
-try:
-    from mmtrack.apis import inference_mot
-    from mmtrack.apis import init_model as init_tracking_model
-    has_mmtrack = True
-except (ImportError, ModuleNotFoundError):
-    has_mmtrack = False
-
-
-def process_mmtracking_results(mmtracking_results):
-    """Process mmtracking results.
-
-    :param mmtracking_results:
-    :return: a list of tracked bounding boxes
-    """
-    person_results = []
-    for track in mmtracking_results['track_results'][0]:
-        person = {}
-        person['track_id'] = int(track[0])
-        person['bbox'] = track[1:]
-        person_results.append(person)
-    return person_results
+                         vis_pose_result)
 
 
 def main():
@@ -35,7 +14,6 @@ def main():
     Using mmdet to detect the human.
     """
     parser = ArgumentParser()
-    parser.add_argument('tracking_config', help='Config file for tracking')
     parser.add_argument('pose_config', help='Config file for pose')
     parser.add_argument('pose_checkpoint', help='Checkpoint file for pose')
     parser.add_argument('--video-path', type=str, help='Video path')
@@ -52,22 +30,11 @@ def main():
     parser.add_argument(
         '--device', default='cuda:0', help='Device used for inference')
     parser.add_argument(
-        '--bbox-thr',
-        type=float,
-        default=0.3,
-        help='Bounding box score threshold')
-    parser.add_argument(
         '--kpt-thr', type=float, default=0.3, help='Keypoint score threshold')
-
-    assert has_mmtrack, 'Please install mmtrack to run the demo.'
 
     args = parser.parse_args()
 
     assert args.show or (args.out_video_root != '')
-    assert args.tracking_config is not None
-
-    tracking_model = init_tracking_model(
-        args.tracking_config, None, device=args.device.lower())
     # build the pose model from a config file and a checkpoint file
     pose_model = init_pose_model(
         args.pose_config, args.pose_checkpoint, device=args.device.lower())
@@ -75,6 +42,11 @@ def main():
     dataset = pose_model.cfg.data['test']['type']
 
     cap = cv2.VideoCapture(args.video_path)
+    assert cap.isOpened(), f'Faild to load video file {args.video_path}'
+
+    fps = cap.get(cv2.CAP_PROP_FPS)
+    size = (int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
+            int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
 
     if args.out_video_root == '':
         save_out_video = False
@@ -83,9 +55,6 @@ def main():
         save_out_video = True
 
     if save_out_video:
-        fps = cap.get(cv2.CAP_PROP_FPS)
-        size = (int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),
-                int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
         fourcc = cv2.VideoWriter_fourcc(*'mp4v')
         videoWriter = cv2.VideoWriter(
             os.path.join(args.out_video_root,
@@ -98,31 +67,26 @@ def main():
     # e.g. use ('backbone', ) to return backbone feature
     output_layer_names = None
 
-    frame_id = 0
     while (cap.isOpened()):
         flag, img = cap.read()
         if not flag:
             break
 
-        mmtracking_results = inference_mot(
-            tracking_model, img, frame_id=frame_id)
-
         # keep the person class bounding boxes.
-        person_results = process_mmtracking_results(mmtracking_results)
+        person_results = [{'bbox': np.array([0, 0, size[0], size[1]])}]
 
         # test a single image, with a list of bboxes.
         pose_results, returned_outputs = inference_top_down_pose_model(
             pose_model,
             img,
             person_results,
-            bbox_thr=args.bbox_thr,
             format='xyxy',
             dataset=dataset,
             return_heatmap=return_heatmap,
             outputs=output_layer_names)
 
         # show the results
-        vis_img = vis_pose_tracking_result(
+        vis_img = vis_pose_result(
             pose_model,
             img,
             pose_results,
@@ -138,8 +102,6 @@ def main():
 
         if cv2.waitKey(1) & 0xFF == ord('q'):
             break
-
-        frame_id += 1
 
     cap.release()
     if save_out_video:
